src/main/java/seedu/workbook/model/WorkBook.java
package seedu.workbook.model;

import static java.util.Objects.requireNonNull;

import java.util.List;

import javafx.collections.ObservableList;
import seedu.workbook.model.person.Person;
import seedu.workbook.model.person.UniquePersonList;

/**
 * Wraps all data at the WorkBook level
 * Duplicates are not allowed (by .isSamePerson comparison)
 */
public class WorkBook implements ReadOnlyWorkBook {

    private final UniquePersonList persons;

    /*
     * The 'unusual' code block below is a non-static initialization block, sometimes used to avoid duplication
     * between constructors. See https://docs.oracle.com/javase/tutorial/java/javaOO/initial.html
     *
     * Note that non-static init blocks are not recommended to use. There are other ways to avoid duplication
     *   among constructors.
     */
    {
        persons = new UniquePersonList();
    }

    public WorkBook() {}

    /**
     * Creates an WorkBook using the Persons in the {@code toBeCopied}
     */
    public WorkBook(ReadOnlyWorkBook toBeCopied) {
        this();
        resetData(toBeCopied);
    }

    //// list overwrite operations

    /**
     * Replaces the contents of the person list with {@code persons}.
     * {@code persons} must not contain duplicate persons.
     */
    public void setPersons(List<Person> persons) {
        this.persons.setPersons(persons);
    }

    /**
     * Resets the existing data of this {@code WorkBook} with {@code newData}.
     */
    public void resetData(ReadOnlyWorkBook newData) {
        requireNonNull(newData);

        setPersons(newData.getPersonList());
    }

    //// person-level operations

    /**
     * Returns true if a person with the same identity as {@code person} exists in the work book.
     */
    public boolean hasPerson(Person person) {
        requireNonNull(person);
        return persons.contains(person);
    }

    /**
     * Adds a person to the work book.
     * The person must not already exist in the work book.
     */
    public void addPerson(Person p) {
        persons.add(p);
    }

    /**
     * Replaces the given person {@code target} in the list with {@code editedPerson}.
     * {@code target} must exist in the work book.
     * The person identity of {@code editedPerson} must not be the same as another existing person in the work book.
     */
    public void setPerson(Person target, Person editedPerson) {
        requireNonNull(editedPerson);

        persons.setPerson(target, editedPerson);
    }

    /**
     * Removes {@code key} from this {@code WorkBook}.
     * {@code key} must exist in the work book.
     */
    public void removePerson(Person key) {
        persons.remove(key);
    }

    //// util methods

    @Override
    public String toString() {
        return persons.asUnmodifiableObservableList().size() + " persons";
        // TODO: refine later
    }

    @Override
    public ObservableList<Person> getPersonList() {
        return persons.asUnmodifiableObservableList();
    }

    @Override
    public boolean equals(Object other) {
        return other == this // short circuit if same object
                || (other instanceof WorkBook // instanceof handles nulls
                && persons.equals(((WorkBook) other).persons));
    }

    @Override
    public int hashCode() {
        return persons.hashCode();
    }
}
